% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsUtil.R
\name{countMissing}
\alias{countMissing}
\title{Count total no. of missing values in a dataframe}
\usage{
countMissing(df)
}
\arguments{
\item{df}{data frame}
}
\description{
sum is na for dataframe
}
\keyword{na,missing}
\keyword{sum}
\keyword{value}
