% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsUtil.R
\name{dfToFactors}
\alias{dfToFactors}
\title{dfToFactors}
\usage{
dfToFactors(ds)
}
\arguments{
\item{ds}{dataframe with numeric, non-categorical data.}
}
\description{
This function allows you to convert all features of a dataframe to factors.
}
\keyword{dataframe}
\keyword{factors}
\keyword{to}
