% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsUtil.R
\name{normalize}
\alias{normalize}
\title{Normalize a column}
\usage{
normalize(x)
}
\arguments{
\item{x}{feature or column of a data frame}
}
\description{
Min-Max normalization for a column.
}
\keyword{minmax,}
\keyword{norm}
