% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsUtil.R
\name{normalizeDF}
\alias{normalizeDF}
\title{Normalize all or selected columns of a dataframe}
\usage{
normalizeDF(df, columnSelector)
}
\arguments{
\item{df}{data frame}

\item{columnSelector}{indexes or range of columns to normalize}
}
\description{
Min-Max normalization for dataframe
}
\keyword{minmax,}
\keyword{norm}
