% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state2Region.R
\name{state2abbrV2}
\alias{state2abbrV2}
\title{Convert state names to abbr, cleans mixed format states}
\usage{
state2abbrV2(statesList)
}
\arguments{
\item{statesList}{vector with state name or abbr}
}
\description{
Convert state names to abbr, cleans mixed format states
}
\keyword{abbreviation}
\keyword{and}
\keyword{names}
\keyword{state}
\keyword{to}
