% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsUtil.R
\name{dfToFactors}
\alias{dfToFactors}
\alias{dfToFactors}
\title{Convert features of a dataframe to factors}
\usage{
dfToFactors(df, col_names)

dfToFactors(df, col_names)
}
\arguments{
\item{df}{dataframe with numeric, non-categorical data.}

\item{col_names}{feature names to convert into factors}

\item{df}{dataframe with numeric, non-categorical data.}
}
\description{
This function allows you to convert all features of a dataframe to factors.

This function allows you to convert a select few features in a dataframe to factors.
}
\examples{
dfToFactors()
dfToFactors()
}
\keyword{factors}
