% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsUtil.R
\name{dfToFactors}
\alias{dfToFactors}
\title{Convert features of a dataframe to factors}
\usage{
dfToFactors(df)
}
\arguments{
\item{df}{dataframe with numeric, non-categorical data.}
}
\description{
This function allows you to convert all features of a dataframe to factors.
}
\examples{
dfToFactors()
}
\keyword{factors}
