% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state2Region.R
\name{groupStatesToRegions}
\alias{groupStatesToRegions}
\title{Convert state name or abbr to region, cleans mixed format states}
\usage{
groupStatesToRegions(states, naLbl)
}
\arguments{
\item{states}{vector with state name or abbr}

\item{naLbl}{if present missing/incorrect values will be assigned this label}
}
\description{
Convert state name or abbr to region, cleans mixed format states
}
\keyword{Region}
\keyword{abbreviation}
\keyword{names}
\keyword{or}
\keyword{state}
\keyword{to}
