% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsUtil.R
\name{listToFactors}
\alias{listToFactors}
\title{listToFactors}
\usage{
listToFactors(df, col_names1)
}
\arguments{
\item{df}{dataframe with numeric, non-categorical data.}

\item{col_names1}{feature names to convert into factors}
}
\description{
This function allows you to convert a select few features in a dataframe to factors.
}
\keyword{factors}
\keyword{to}
\keyword{vector}
