% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsUtil.R
\name{normalizeDF}
\alias{normalizeDF}
\title{Normalize all or selected columns of a dataframe}
\usage{
normalizeDF(df, columns)
}
\arguments{
\item{df}{data frame}
}
\description{
Min-Max normalization for dataframe
}
\examples{
normalizeDF()
}
\keyword{minmax,}
\keyword{norm}
